Fix tests for elections-phragmen
